renamed routePathListObject -> routePathListItem
