+++ src/components/sidebar/routePathView/routePathListTab/routePathListItem.scss
@@ -18,7 +18,7 @@
     border-bottom: 1.5px solid $lightGrey;
     position: relative;
 
-    .headerIcon {
+    .listIcon {
         width: 30px;
         position: absolute;
         height: 30px;
